+++ nomen/src/subcommands/index/events/event_data.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 use anyhow::bail;
-use nomen_core::util::{EventExtractor, Hash160, Name, Nsid};
+use nomen_core::{EventExtractor, Hash160, Name, Nsid};
 use nostr_sdk::{Event, EventId};
 use secp256k1::XOnlyPublicKey;
 
@@ -55,7 +55,7 @@ impl EventData {
 
 #[cfg(test)]
 mod tests {
-    use nomen_core::util::Nsid;
+    use nomen_core::Nsid;
 
     use super::*;
