+++ nomen_core/src/kind.rs
